Added email_verified_at to user. Changed primary key field from integer to bigInteger in migrations.

+++ database/migrations/2018_04_26_141254_create_reviews_table.php
@@ -20,8 +20,8 @@ class CreateReviewsTable extends Migration
     public function up() :void
     {
         Schema::create('reviews', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('restaurant_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('restaurant_id')->unsigned();
             $table->text('comment');
             $table->integer('rating');
             $table->timestamps();
